+++ ton-http-api/pyTON/__main__.py
@@ -62,3 +62,8 @@ def main():
 
     from pyTON.main import app
     uvicorn.run(app, host=args.host, port=args.port, log_level=args.logs_level.lower())
+
+
+# run with python3 -m pyTON
+if __name__ == '__main__':
+    main()
