+++ ton-http-api/pyTON/core/cache.py
@@ -2,7 +2,7 @@ import aioredis
 import ring
 
 from ring.func.asyncio import Aioredis2Storage
-from pyTON.settings import RedisCacheSettings
+from pyTON.core.settings import RedisCacheSettings
 
 
 class TonlibResultRedisStorage(Aioredis2Storage):
@@ -30,7 +30,6 @@ class RedisCacheManager:
     def __init__(self, cache_settings: RedisCacheSettings):
         self.cache_settings = cache_settings
         self.cache_redis = aioredis.from_url(f"redis://{cache_settings.redis.endpoint}:{cache_settings.redis.port}")
-        print(f"WARNING! {self.cache_redis}")
 
     def cached(self, expire=0, check_error=True):
         storage_class = TonlibResultRedisStorage if check_error else Aioredis2Storage
