+++ ton-http-api/pyTON/core/tonlib/worker.py
@@ -1,20 +1,17 @@
 import asyncio
-import traceback
 import random
 import time
 import queue
 import multiprocessing as mp
 
-from pyTON.settings import TonlibSettings
-from pyTON.models import TonlibWorkerMsgType, TonlibClientResult
-from pytonlib import TonlibClient, TonlibException
 from datetime import datetime
 from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
+from typing import Optional
 
-from enum import Enum
-from dataclasses import dataclass
-from typing import Any, Dict, Optional
+from pyTON.core.settings import TonlibSettings
+from pyTON.core.tonlib.models import TonlibWorkerMsgType, TonlibClientResult
+from pytonlib import TonlibClient, TonlibException
 
 from loguru import logger
 
@@ -107,7 +104,7 @@ class TonlibWorker(mp.Process):
                 last_block = masterchain_info["last"]["seqno"]
                 self.timeout_count = 0
             except TonlibException as e:
-                logger.error("TonlibWorker #{ls_index:03d} report_last_block exception of type {exc_type}: {exc}", ls_index=self.ls_index, exc_type=type(e).__name__, exc=e)
+                logger.debug("TonlibWorker #{ls_index:03d} report_last_block exception of type {exc_type}: {exc}", ls_index=self.ls_index, exc_type=type(e).__name__, exc=e)
                 self.timeout_count += 1
 
             if self.timeout_count >= 10:
@@ -124,7 +121,7 @@ class TonlibWorker(mp.Process):
                 block_transactions = await self.tonlib.get_block_transactions(-1, -9223372036854775808, random.randint(2, 4096), count=10)
                 is_archival = block_transactions.get("@type", "") == "blocks.transactions"
             except TonlibException as e:
-                logger.error("TonlibWorker #{ls_index:03d} report_archival exception of type {exc_type}: {exc}", ls_index=self.ls_index, exc_type=type(e).__name__, exc=e)
+                logger.debug("TonlibWorker #{ls_index:03d} report_archival exception of type {exc_type}: {exc}", ls_index=self.ls_index, exc_type=type(e).__name__, exc=e)
             self.is_archival = is_archival
             await self.loop.run_in_executor(self.threadpool_executor, self.output_queue.put, (TonlibWorkerMsgType.ARCHIVAL_UPDATE, self.is_archival))
             await asyncio.sleep(600)
