+++ ton-http-api/pyTON/schemas/__init__.py
@@ -1,0 +1,22 @@
+from .http import (
+    TonRequestJsonRPC,
+    TonResponse,
+    TonResponseGeneric,
+    TonResponseJsonRPC
+)
+from .ton import (
+    BlockId,
+    BlockHeader,
+    SmartContract,
+    AdressUserFriendly,
+    AddressForms,
+    MasterchainInfo,
+    ExternalMessage,
+    SerializedBoc,
+    MsgDataRaw,
+    Message,
+    TransactionId,
+    Transaction,
+    check_tonlib_type,
+    address_state
+)
