+++ ton-http-api/pyTON/schemas/http.py
@@ -1,0 +1,32 @@
+from typing import Optional, TypeVar, Union
+from pydantic.generics import GenericModel, Generic
+from pydantic import BaseModel
+
+
+ResultT = TypeVar('ResultT')
+
+
+class TonResponseGeneric(GenericModel, Generic[ResultT]):
+    ok: bool
+    result: Optional[ResultT]
+    error: Optional[str] = None
+    code: Optional[int] = None
+
+
+class TonResponse(TonResponseGeneric[Union[str, list, dict, None]]):
+    pass
+
+
+class TonResponseJsonRPC(BaseModel):
+    id: str
+    jsonrpc: str = "2.0"
+    result: Optional[ResultT]
+    error: Optional[str] = None
+    code: Optional[int] = None
+
+
+class TonRequestJsonRPC(BaseModel):
+    method: str
+    params: dict = {}
+    id: Optional[str] = None
+    jsonrpc: Optional[str] = None
